+++ src/widgets/slider_widget.rs
@@ -19,11 +19,13 @@ use crate::render::widget_cache::WidgetContainer;
 use crate::render::widget_config::*;
 use crate::render::{Points, Size, POINT_X, POINT_Y, SIZE_HEIGHT, SIZE_WIDTH};
 
-use sdl2::render::Canvas;
+use sdl2::render::{Canvas, Texture};
 use sdl2::video::Window;
 
 use crate::render::canvas_helper::CanvasHelper;
 use crate::render::layout_cache::LayoutContainer;
+use crate::render::texture_cache::TextureCache;
+use crate::render::texture_store::TextureStore;
 use crate::widgets::slider_widget::SliderOrientation::{SliderHorizontal, SliderVertical};
 use sdl2::pixels::Color;
 use sdl2::rect::{Point, Rect};
@@ -50,6 +52,7 @@ pub struct SliderWidget {
     config: WidgetConfig,
     system_properties: HashMap<i32, String>,
     callback_registry: CallbackRegistry,
+    _texture_store: TextureStore,
     min: u32,
     max: u32,
     current: u32,
@@ -77,6 +80,7 @@ impl SliderWidget {
             config: WidgetConfig::new(points, size),
             system_properties: HashMap::new(),
             callback_registry: CallbackRegistry::new(),
+            _texture_store: TextureStore::default(),
             min,
             max,
             current,
@@ -114,7 +118,7 @@ impl CanvasHelper for SliderWidget {}
 /// This is the `Widget` implementation of the `SliderWidget`.
 impl Widget for SliderWidget {
     /// Draws the `SliderWidget` contents.
-    fn draw(&mut self, c: &mut Canvas<Window>) {
+    fn draw(&mut self, c: &mut Canvas<Window>, _t: &mut TextureCache) -> Option<&Texture> {
         if self.orientation == SliderHorizontal {
             let base_color = self.get_color(CONFIG_COLOR_BASE);
 
@@ -266,6 +270,8 @@ impl Widget for SliderWidget {
         }
 
         self.draw_bounding_box(c);
+
+        None
     }
 
     /// When a mouse enters the bounds of the `Widget`, this function is triggered.