+++ src/widgets/text_widget.rs
@@ -20,10 +20,12 @@ use crate::render::widget_cache::WidgetContainer;
 use crate::render::widget_config::*;
 use crate::render::{Points, Size};
 
-use sdl2::render::{Canvas, TextureQuery};
+use sdl2::render::{Canvas, Texture, TextureQuery};
 use sdl2::ttf::FontStyle;
 use sdl2::video::Window;
 
+use crate::render::texture_cache::TextureCache;
+use crate::render::texture_store::TextureStore;
 use sdl2::rect::Rect;
 use std::any::Any;
 use std::collections::HashMap;
@@ -48,6 +50,7 @@ pub struct TextWidget {
     config: WidgetConfig,
     system_properties: HashMap<i32, String>,
     callback_registry: CallbackRegistry,
+    _texture_store: TextureStore,
     font_name: String,
     font_style: FontStyle,
     font_size: i32,
@@ -74,6 +77,7 @@ impl TextWidget {
             config: WidgetConfig::new(points, size),
             system_properties: HashMap::new(),
             callback_registry: CallbackRegistry::new(),
+            _texture_store: TextureStore::default(),
             font_name,
             font_style,
             font_size,
@@ -98,7 +102,7 @@ impl TextWidget {
 /// copied to the canvas after rendering.  It uses blended mode texture mapping, which may be slow (as
 /// described by the SDL2 documentation), so this might change later to use 8 bit color mapping.
 impl Widget for TextWidget {
-    fn draw(&mut self, c: &mut Canvas<Window>) {
+    fn draw(&mut self, c: &mut Canvas<Window>, _t: &mut TextureCache) -> Option<&Texture> {
         let base_color = self.get_color(CONFIG_COLOR_BASE);
         let text_max_width =
             self.get_size(CONFIG_SIZE)[0] - ((self.get_numeric(CONFIG_BORDER_WIDTH) * 2) as u32);
@@ -143,6 +147,8 @@ impl Widget for TextWidget {
             Rect::new(texture_x, texture_y, width, height),
         )
         .unwrap();
+
+        None
     }
 
     /// Monitors for changes in the text, color changes, or font sizes.